refactor(lib): split memory fs to remove max-lines disables
